+++ scripts/doc8.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+doc8 README.rst
+doc8 CHANGELOG.rst
+doc8 docs/